Realizado el sistema premium y retocada la forma de eliminar y modificar grupos, comenzado el despliegue de gráficos

+++ AppChat/src/main/java/modelo/Descuento.java
@@ -1,5 +1,5 @@
 package modelo;
 
 public interface Descuento {
-	public int getDescuento();
+	public double getDescuento();
 }
